+++ tox_server/exc.py
@@ -1,0 +1,7 @@
+__all__ = ["TSException"]
+
+
+class TSException(Exception):
+    """Base class for all tox-server exceptions"""
+
+    pass
